+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 | Difficulty | Resolved | Total |
 | :--------- | :------- | :---- |
 | Easy       | 21       | 492   |
-| Medium     | 49       | 837   |
+| Medium     | 50       | 837   |
 | Hard       | 0        | 328   |
 
 ## 0005 - Longest palindromic substring
@@ -315,6 +315,11 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 - Deep First Search
 - Not remove number `0`: `010010` -> `[0.10.0.10, 0.100.1.0]`
 - [Golang]: `strconv.ParseUint`, `strconv.FormatUint`
+
+## 0094 - Binary tree inorder traversal
+
+- Deep First Search
+- Recursion
 
 ## 0415 - Add strings